+++ src/main/java/ca/corefacility/bioinformatics/irida/model/assembly/GenomeAssembly.java
@@ -125,6 +125,11 @@ public abstract class GenomeAssembly extends IridaResourceSupport
 	public abstract Path getFile();
 
 	@Override
+	public String getFileName() {
+		return getFile().getFileName().toString();
+	}
+
+	@Override
 	public int hashCode() {
 		return Objects.hash(id, createdDate);
 	}